fix: `verification_method` is not required for DID Document
Re. https://www.w3.org/TR/did-core/#did-document-properties, the
veriication_method property is not required. In fact, the TBD-hosted
DID document used is tests does not include the Verification method.

Make it optional, especially during (de-)serialization to ensure
compatibility with the spec.

Note: the web5-spec disagrees with the DID-core spec on this, it may
be a legacy but it's possible the web5-spec should be updated as well
to match did-core.