fix: disable jsonschema resolution to support web5-rs in WASM builds
Without disabling the resolve-http feature, the jsonschema crate fails to build
on WASM environments due to use of `reqwest::blocking`, which is not built in
WASM environments. Stranger6667/jsonschema-rs#222 has additional context.

The use of jsonschema, currently at least, does not use resolution and is used
against a schema defined in code for Verifiable Credentials